Move SoX filter test to test_functional_filtering
